sms/interceptor.https.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/samuelgoto/sms-receiver">
<title>Tests the SMS Receiver API</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!--
 sms_provider.js is a testing framework that enables engines to test the sms
 receiver API by intercepting the connection between the browser and the
 underlying operating system and mock its behavior.

 Usage:

 1) Include <script src="./sms_provider.js"></script> in your test.
 2) Set expectations
  await expect(getNextMessage).andReturn((timeout) => {
    // mock behavior
  })
 3) Call navigator.sms.receive()
 4) Verify results

 The mocking API is browser agnostic and is designed such that other engines
 could implement it too.

 Here are the symbols that are exposed to tests that need to be implemented
 per engine:

 - function getNextMessage(): the main/only function that can be mocked.
 - function expect(): the main/only function that enables us to mock it.
 - enum State {kSuccess, kTimeout}: allows you to mock success/failures.

  -->
<script src="./sms_provider.js"></script>
<script>
'use strict';

promise_test(async t => {
  await expect(getNextMessage).andReturn((timeout) => {
      return Promise.resolve({
        sms: {
          content: "hello",
          status: Status.kSuccess,
        }
      });
  });

  let sms = await navigator.sms.receive();

  assert_equals(sms.content, "hello");
}, 'Basic usage');

promise_test(async t => {
  await expect(getNextMessage).andReturn((timeout) => {
      return Promise.resolve({
        sms: {
          content: "",
          status: Status.kTimeout,
        }
      });
  });

  let error = await new Promise(function(resolve) {
    navigator.sms.receive().catch((error) => {
      resolve(error);
    });
  });
  assert_equals(error.name, "TimeoutError");
  assert_equals(error.message, "SMSReceiver timed out.");
}, 'Deal with timeouts');

promise_test(async t => {
  let error = await new Promise(function(resolve) {
  navigator.sms.receive({timeout: 0}).catch((error) => {
      resolve(error);
    });
  });
  assert_equals(error.name, "NotSupportedError");
  assert_equals(error.message, "Invalid timeout.");
}, 'Should throw error with invalid timeout (0)');

promise_test(async t => {
  let error = await new Promise(function(resolve) {
  navigator.sms.receive({timeout: null}).catch((error) => {
      resolve(error);
    });
  });
  assert_equals(error.name, "NotSupportedError");
  assert_equals(error.message, "Invalid timeout.");
}, 'Should throw error with invalid timeout (null)');

promise_test(async t => {
  let error = await new Promise(function(resolve) {
  navigator.sms.receive({timeout: -1}).catch((error) => {
      resolve(error);
    });
  });
  assert_equals(error.name, "NotSupportedError");
  assert_equals(error.message, "Invalid timeout.");
}, 'Should throw error with invalid timeout (-1)');

promise_test(async t => {
  let error = await new Promise(function(resolve) {
  navigator.sms.receive({timeout: NaN}).catch((error) => {
      resolve(error);
    });
  });
  assert_equals(error.name, "NotSupportedError");
  assert_equals(error.message, "Invalid timeout.");
}, 'Should throw error with invalid timeout (NaN)');

promise_test(async t => {
  await expect(getNextMessage).andReturn((timeout) => {
      return Promise.resolve({
        sms: {
          content: "hello",
          status: Status.kSuccess,
        }
      });
  });

  let sms = await navigator.sms.receive({timeout: undefined});
  assert_equals(sms.content, "hello");
}, 'Should use default value for timeout (undefined)');

</script>
